app_clock: promote RTC shell to top-level `clock` command
Move the RTC shell commands out of the debug `tester clock` subcommand into a
standalone top-level `clock get/set/sync`, owned by the app_clock module itself.

- app_clock.c: add the get/set/sync handlers + `SHELL_CMD_REGISTER(clock, …)`
  under a CONFIG_SHELL guard (logic unchanged, moved verbatim).
- app_tester.c: drop the clock handlers, sub_clock set, the `tester clock`
  registration, and the now-unused app_clock.h / <time.h> includes.

Verified on hardware (debug build): `clock` lists get/set/sync; `clock get`
reports "RTC not set yet" then `clock set <unix>` / `clock get` round-trips;
`clock sync` queues a DeviceTimeReq. Release build compiles (shell compiled out).
